First pass through SSE1 instructions.
git-svn-id: svn://svn.valgrind.org/vex/trunk@1177

+++ VEX/priv/guest-amd64/gdefs.h
@@ -117,8 +117,8 @@ extern ULong amd64g_create_mxcsr ( ULong sseround );
 
 extern ULong amd64g_calculate_mmx_pmaddwd  ( ULong, ULong );
 extern ULong amd64g_calculate_mmx_psadbw   ( ULong, ULong );
-extern UInt  amd64g_calculate_mmx_pmovmskb ( ULong );
-extern UInt  amd64g_calculate_sse_pmovmskb ( ULong w64hi, ULong w64lo );
+extern ULong amd64g_calculate_mmx_pmovmskb ( ULong );
+extern ULong amd64g_calculate_sse_pmovmskb ( ULong w64hi, ULong w64lo );
 
 
 /* --- DIRTY HELPERS --- */
